Adds new feature type for offsets with an Xpath implementation.
This also adds a little more convenience for default feature types.
Namely, if a special prefix is used (e.g., `#` or `@`) and the feature
has not been added to an FC yet, then a default feature type is chosen
based on the prefix.

`StringCounter` remains the de facto default, but `XpathFeatureOffsets`
is now the default for `@` based features.